Fixing a crash when phone connection is spotty
The proximate cause was a state transition of phone not connected -> phone
connected triggering a sync() while the app was still in an init state. This
caused a crash b/c mHueClient would have been null.

The root cause was treating PHONE NOT CONNECTED as a proper state. Really it's
more of an exception, not part of the app flow.